src/tactical_empathy/requirements.txt
django==4.2.11
djangorestframework==3.14.0
drf-spectacular==0.27.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-cors-headers==4.3.1
pydantic==2.5.0
whitenoise==6.6.0
openai
gunicorn==21.2.0